tweaks to script to copy RT tickets to github
take dry run opts, as well as options to specify the dist, ticket
id on the command line. split logic into subroutines to clean up
the script a little. comment on copied tickets in RT to make it
clear that they have been copied from RT to github and any future
correspondence should be taken there